moves TestConnectionGating to new file
